lib: remove lua dep and fix type conversion warning

+++ lib/generic/array.h
@@ -132,7 +132,7 @@ static inline void array_std_free(void *baton, void *p)
  */
 #define array_push_mm(array, val, reserve, baton) \
 	(int)((array).len < (array).cap ? ((array).at[(array).len] = val, (array).len++) \
-		: (array_reserve_mm(array, ((array).cap + 1), reserve, baton) < 0 ? -1 \
+		: (array_reserve_mm(array, ((array).cap + 1), reserve, baton) < 0 ? (size_t)-1 \
 			: ((array).at[(array).len] = val, (array).len++)))
 
 /**
